Build all the images from a set of SVG files.
The SVG files are split up so that there is no shape duplication.

Use rsvg-convert since inkscape is fat and slow and doesn't seem to
support filter chaining in the way that we need it.

Also include copyright, license and source code in SVG format for the
KDE and Neverball logos.

Pre-process the KDE and Neverball logos before including them in the
icons so that we can swap them out or change effects more easily.

Signed-off-by: ANT <ant007h@gmail.com>